[0140] 将 cycles/shared-info 环检测机制从 s7.c 迁移到 s7_scheme_write.c
