+++ devel/0140.md
@@ -64,6 +64,31 @@ object->port 区块本身是内部打印引擎（77 个 `*_to_port` 函数，无
 - pos_int_to_str 系列等小工具可随迁
 - eq/equal 区块（33975）仅用 clear_shared_info，导出声明即可
 
+## 2026-08-26 第二步：迁移 cycles/shared-info 机制（第一步提交）
+
+### What
+1. 将 s7.c 的 cycles 区块（shared_ref / flip_ref / peek_shared_ref_1 / peek_shared_ref /
+   enlarge_shared_info / check_collected / collect_shared_info / collect_vector_info /
+   make_shared_info / free_shared_info / clear_shared_info / load_shared_info /
+   cyclic_sequences_p_p，约 443 行）迁移到 s7_scheme_write.c
+2. `pair_append`（通用 pair 工具，eval 用 21 处）从 object->port 区块挪出，去掉 static 导出
+3. `hash_keys_not_cyclic`（s7.c 哈希表区块）去掉 static，声明加入 s7_internal_helpers.h
+4. `shared_info_t` 在两个定义处（s7.c / s7_internal.h）都加 `struct shared_info` 标签，
+   s7_scheme_write.h 用 `struct shared_info *` 前置声明导出走写模块接口
+5. s7_internal.h 新增镜像宏 42 个（is_collected_unchecked / set_cyclic / has_structure /
+   t_structure_p 等类型谓词表 extern 声明等），Malloc/Calloc/Realloc 带 S7_DEBUGGING 条件分支
+6. write.c 改用已有的 `s7i_check_free_heap_size` 桥接
+
+### How
+- 编译错误驱动的固定点迭代：gcc -fsyntax-only 收集未定义标识符，从 s7.c 提取对应
+  `#define` 块（含续行）追加到 s7_internal.h，直到收敛
+- s7_scheme_write.c 改为 include s7_internal.h（含完整 struct s7_scheme 与宏）
+
+### 验证
+- xmake b goldfish 编译通过
+- bin/gf test --all：1515 个测试全部通过
+- 所有 include s7_internal.h 的模块（write/read/let/continuation）语法检查通过
+
 ## 2026-08-26 第一步：为涉及函数补齐测试与文档
 
 ### What
