+++ src/s7_scheme_write.h
@@ -11,9 +11,26 @@
 
 #include "s7.h"
 
+#include <stdbool.h>
+#include <stdint.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
+
+/* cycles / shared-info: circular reference detection for the printer */
+struct shared_info;
+int32_t shared_ref(struct shared_info *ci, s7_pointer p);
+void flip_ref(struct shared_info *ci, s7_pointer p);
+int32_t peek_shared_ref_1(struct shared_info *ci, s7_pointer p);
+int32_t peek_shared_ref(struct shared_info *ci, s7_pointer p);
+void enlarge_shared_info(struct shared_info *ci);
+struct shared_info *make_shared_info(s7_scheme *sc);
+void free_shared_info(struct shared_info *ci);
+struct shared_info *clear_shared_info(struct shared_info *ci);
+struct shared_info *load_shared_info(s7_scheme *sc, s7_pointer top, bool stop_at_print_length, struct shared_info *ci);
+bool collect_shared_info(s7_scheme *sc, struct shared_info *ci, s7_pointer top, bool stop_at_print_length);
+s7_pointer cyclic_sequences_p_p(s7_scheme *sc, s7_pointer obj);
 
 /* Public API implementations */
 void s7_newline(s7_scheme *sc, s7_pointer port);
